requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn


# Google genai 
langchain-google-genai


# Groq Integration
langchain-groq

pypdf

# Vector Stores
chromadb
langchain_chroma


# Vector Store Utilities
tf-keras


# mistral ai
langchain-mistralai